Remove separate merge_single_qubit_gates_to_phxz_perf benchmark
The benchmarks in merge_single_qubit_gates_to_phxz are already a portion
of the existing dynamical_decoupling benchmark so it is not worth having
a separate benchmark.
